refactoring create contact screen
