+++ src/components/DateButton.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { TouchableOpacity } from 'react-native-gesture-handler';
+import { Keyboard } from 'react-native';
+
+const DateButton = ({ dateFn, text }) => (
+  <TouchableOpacity
+    onPress={() => {
+      Keyboard.dismiss();
+      dateFn();
+    }}
+  >
+    {text}
+  </TouchableOpacity>
+);
+
+export default DateButton;
